requirements.txt
agents
beautifulsoup4
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
openai
pandas
python-dotenv
pyyaml
requests
searchconsole
